feat(patient.service.ts):implemented func for family member operations: BATCH

+++ src/patient/services/patient.service.ts
@@ -1,7 +1,9 @@
 import { HttpStatus, Injectable, Logger } from '@nestjs/common';
 import { Patient } from '../schemas/patient.schema';
 import {
+  CreateFamilyMemberType,
   CreatePatientType,
+  UpdateFamilyMemberType,
   UpdatePatientProfileType,
 } from '../interface/patient.interface';
 import { ErrorCodes, PatientError } from 'src/shared';
@@ -10,6 +12,10 @@ import { Model, MongooseError } from 'mongoose';
 import { PatientDao } from '../dao/patient.dao';
 import { PatientGuard } from '../guards/patient.guard';
 
+/**
+ * @file: Patient Service
+ * @author: 3illbaby
+ */
 @Injectable()
 export class PatientService {
   constructor(
@@ -63,6 +69,177 @@ export class PatientService {
     } catch (error) {
       Logger.error(error);
       throw new PatientError('An error occurred while creating patient');
+    }
+  }
+
+  async addFamilyMember(args: {
+    walletAddress: string;
+    familyMember: CreateFamilyMemberType;
+  }) {
+    const { walletAddress, familyMember } = args;
+    const {
+      id,
+      name,
+      relationship,
+      email,
+      address,
+      age,
+      dob,
+      bloodGroup,
+      genotype,
+    } = familyMember;
+    try {
+      const patient =
+        await this.patientDao.fetchPatientByAddress(walletAddress);
+
+      if (!patient) {
+        return {
+          success: HttpStatus.NOT_FOUND,
+          message: 'patient not found',
+        };
+      }
+
+      const sanitizeRelationship = relationship.toLowerCase();
+      const sanitizedArgs = {
+        id,
+        principalPatient: walletAddress,
+        name,
+        relationship: sanitizeRelationship,
+        email,
+        address,
+        age,
+        dob,
+        bloodGroup,
+        genotype,
+      };
+      const familyMemberExist = patient.familyMembers.find(
+        (member) => member.id === id,
+      );
+
+      if (familyMemberExist) {
+        return {
+          success: HttpStatus.CONFLICT,
+          message: 'family member already exist',
+        };
+      }
+
+      const newFamilyMember =
+        await this.patientDao.returnFamilyMembers(sanitizedArgs);
+      patient.familyMembers.push(newFamilyMember);
+      await patient.save();
+      return {
+        success: HttpStatus.OK,
+        message: 'Family member added successfully',
+      };
+    } catch (error) {
+      console.error(error);
+      throw new PatientError('An error occurred while adding family member');
+    }
+  }
+
+  async listFamilyMember(walletAddress: string) {
+    try {
+      const patient =
+        await this.patientDao.fetchPatientByAddress(walletAddress);
+      if (!patient) {
+        return {
+          success: HttpStatus.NOT_FOUND,
+          message: 'Patient not found',
+        };
+      }
+      const familyMembers = patient.familyMembers;
+      if (!familyMembers) {
+        return {
+          success: HttpStatus.FOUND,
+          members: [],
+          message: 'No family members added',
+        };
+      }
+      return {
+        success: HttpStatus.FOUND,
+        members: familyMembers,
+        message: 'Family members found',
+      };
+    } catch (error) {
+      console.error(error);
+      throw new Error('An error occurred while listing family member');
+    }
+  }
+
+  async getFamilyMemberById(args: { walletAddress: string; memberId: number }) {
+    const { walletAddress, memberId } = args;
+    try {
+      const patient =
+        await this.patientDao.fetchPatientByAddress(walletAddress);
+      if (!patient) {
+        return {
+          success: HttpStatus.NOT_FOUND,
+          message: 'Patient not found',
+        };
+      }
+      const familyMember = patient.familyMembers;
+      const member = familyMember.find((member) => member.id === memberId);
+      if (!member) {
+        return {
+          success: HttpStatus.NOT_FOUND,
+          message: 'Member not found',
+        };
+      }
+
+      return {
+        success: HttpStatus.OK,
+        member,
+      };
+    } catch (error) {
+      console.error(error);
+      throw new PatientError('an error occurred while fetching family member');
+    }
+  }
+
+  async editFamilyMember(args: {
+    walletAddress: string;
+    familyMemberId: number;
+    updateData: UpdateFamilyMemberType;
+  }) {
+    const { walletAddress, familyMemberId, updateData } = args;
+    try {
+      const patientExist =
+        await this.patientGuard.validatePatient(walletAddress);
+      if (!patientExist) {
+        return {
+          success: HttpStatus.NOT_FOUND,
+          message: 'Patient not found',
+        };
+      }
+
+      const patient =
+        await this.patientDao.fetchPatientByAddress(walletAddress);
+
+      const familyMemberExists = patient.familyMembers.find(
+        (member) => member.id === familyMemberId,
+      );
+
+      if (!familyMemberExists) {
+        return {
+          success: HttpStatus.NOT_FOUND,
+          message: 'Family member not found',
+        };
+      }
+
+      const familyMember = await this.patientDao.updateFamilyMember(
+        walletAddress,
+        familyMemberId,
+        updateData,
+      );
+
+      return {
+        success: HttpStatus.OK,
+        message: 'Family member updated successfully',
+        familyMember,
+      };
+    } catch (error) {
+      console.error(error);
+      throw new PatientError('An error occurred while editing family member');
     }
   }
 
@@ -120,6 +297,23 @@ export class PatientService {
   }
 
   async deletePatientByAddress(walletAddress: string) {
-    return await this.patientModel.deleteOne({ walletAddress });
+    try {
+      const patientExists =
+        await this.patientGuard.validatePatient(walletAddress);
+      if (!patientExists) {
+        return {
+          success: HttpStatus.NOT_FOUND,
+          message: 'Patient not found',
+        };
+      }
+      await this.patientModel.deleteOne({ walletAddress });
+      return {
+        success: HttpStatus.OK,
+        message: 'Patient deleted successfully',
+      };
+    } catch (error) {
+      console.error(error);
+      throw new PatientError('An error occurred while deleting patient');
+    }
   }
 }
